% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_map}
\alias{cell_map}
\title{cell_map}
\format{A SpatialPolygonsDataFrame with 2148 features:
\describe{
  \item{\code{OBJECTID}}{Unique row identifier}
  \item{\code{Prov_ID}}{Unique province identifier}
  \item{\code{Prov_Name}}{Name of province}
  \item{\code{Dist_ID}}{Unique district identifier}
  \item{\code{District}}{Name of district}
  \item{\code{Sect_ID}}{Unique sector identifier}
  \item{\code{Sector}}{Name of sector}
  \item{\code{Cell_ID}}{Unique cell identifier}
  \item{\code{Cell}}{Name of cell}
}}
\source{
National Instituate of Statistics of Rwanda
    \url{http://geodata-nisr-rwanda.opendata.arcgis.com}
}
\usage{
cell_map
}
\description{
Aministrative level 4 (cell) map of Rwanda
}
\keyword{datasets}
