% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_map}
\alias{district_map}
\title{district_map}
\format{A SpatialPolygonsDataFrame with 30 features:
\describe{
  \item{\code{OBJECTID}}{Unique row identifier}
  \item{\code{Prov_ID}}{Unique province identifier}
  \item{\code{Prov_Name}}{Name of province}
  \item{\code{Dist_ID}}{Unique district identifier}
  \item{\code{District}}{Name of district}
}}
\source{
National Instituate of Statistics of Rwanda
    \url{http://geodata-nisr-rwanda.opendata.arcgis.com}
}
\usage{
district_map
}
\description{
Aministrative level 2 (district) map of Rwanda
}
\keyword{datasets}
