% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_map}
\alias{province_map}
\title{province_map}
\format{A SpatialPolygonsDataFrame with 5 features:
\describe{
  \item{\code{OBJECTID}}{Unique row identifier}
  \item{\code{Prov_ID}}{Unique province identifier}
  \item{\code{Prov_Name}}{Name of province}
  \item{\code{Province}}{Name of province (Kinyarwanda)}
}}
\source{
National Instituate of Statistics of Rwanda
    \url{http://geodata-nisr-rwanda.opendata.arcgis.com}
}
\usage{
province_map
}
\description{
Aministrative level 1 (province) map of Rwanda
}
\keyword{datasets}
