% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sector_map}
\alias{sector_map}
\title{sector_map}
\format{A SpatialPolygonsDataFrame with 416 features:
\describe{
  \item{\code{OBJECTID}}{Unique row identifier}
  \item{\code{Prov_ID}}{Unique province identifier}
  \item{\code{Prov_Name}}{Name of province}
  \item{\code{Dist_ID}}{Unique district identifier}
  \item{\code{District}}{Name of district}
  \item{\code{Sect_ID}}{Unique sector identifier}
  \item{\code{Sector}}{Name of sector}
}}
\source{
National Instituate of Statistics of Rwanda
    \url{http://geodata-nisr-rwanda.opendata.arcgis.com}
}
\usage{
sector_map
}
\description{
Aministrative level 3 (sector) map of Rwanda
}
\keyword{datasets}
