% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villages}
\alias{villages}
\title{villages}
\format{A data frame containing 12 columns and 14816 rows:
\describe{
\item{\code{Prov_ID}}{Province ID}
\item{\code{Province}}{Name of Province}
\item{\code{Dist_ID}}{District ID}
\item{\code{District}}{Name of District}
\item{\code{Sect_ID}}{Sector ID}
\item{\code{Sector}}{Name of Sector}
\item{\code{Cell_ID}}{Cell ID}
\item{\code{Cell}}{Name of Cell}
\item{\code{Vill_ID}}{Village ID}
\item{\code{Village}}{Name of Village}
\item{\code{Status}}{Whether village is rural or urban}
\item{\code{FID}}{FID}
}}
\source{
National Instituate of Statistics of Rwanda
    \url{http://geodata-nisr-rwanda.opendata.arcgis.com}
}
\usage{
villages
}
\description{
Villages dataset of Rwanda
}
\keyword{datasets}
